FROMLIST: clk / soc: mediatek: Move mt8173 MMSYS to platform driver
There is no strong reason for this to use CLK_OF_DECLARE instead of
being a platform driver. Plus, MMSYS provides clocks but also a shared
register space for the mediatek-drm and the mediatek-mdp
driver. So move the MMSYS clocks to a new platform driver and also
create a new MMSYS platform driver in drivers/soc/mediatek that
instantiates the clock driver.

Signed-off-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: CK Hu <ck.hu@mediatek.com>
Series-changes: 12
- Leave the clocks part in drivers/clk (clk-mt8173-mm)
- Instantiate the clock driver from the mtk-mmsys driver.
- Add default config option to not break anything.
- Removed the Reviewed-by CK tag as changed the organization.

Series-changes: 10
- Renamed to be generic mtk-mmsys
- Add driver data support to be able to support diferent SoCs

Series-changes: 9
- Move mmsys to drivers/soc/mediatek (CK)

Series-changes: 8
- Be a builtin_platform_driver like other mediatek mmsys drivers.

Series-changes: 7
- Free clk_data->clks as well
- Get rid of private data structure