FROMLIST: dt-bindings: power: Add bindings for the Mediatek SCPSYS power domains controller
The System Control Processor System (SCPSYS) has several power management
related tasks in the system. Add the bindings to define the power
domains for the SCPSYS power controller.

Co-developed-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Series-to: lkml
Series-cc: mediatek, collabora, weiyi.lu@mediatek.com, fparent@baylibre.com
Series-version: 3
Series-changes: 3
- Use hex for unit-addresses.
- Define child nodes for nested power domains even are duplicated, but
  more clear than adding a regex scaped to be a valid URI.

Series-changes: 2
- Use generic node names (power-domain).
- Define valid values for common properties like #power-domain-cells.

Cover-letter:
soc: mediatek: pm-domains: Add new driver for SCPSYS power domains controller
Dear all,

This is a new driver with the aim to deprecate the mtk-scpsys driver.
The problem with that driver is that, in order to support more Mediatek
SoCs you need to add some logic to handle properly the power-up
sequence of newer Mediatek SoCs, doesn't handle parent-child power
domains and need to hardcode all the clocks in the driver itself. The
result is that the driver is getting bigger and bigger every time a
new SoC needs to be supported.

All this information can be getted from a properly defined binding, so
can be cleaner and smaller, hence, we implemented a new driver. For
now, only MT8173 and MT8183 is supported but should be fairly easy to
add support for new SoCs.

Two important notes:

1. Support for MT8183 is not ready to land yet because has some
   dependencies, i.e mmsys support is still missing.

2. Support for MT8192. I picked the patches [1] from Weiyi Lu and
   adapted to this new series. I posted only for reference due that this
   new version has some changes that affects that patchset.

Only patches from 1 to 9 are ready, the others are provided for reference and test.

[1] https://patchwork.kernel.org/project/linux-mediatek/list/?series=368821

Best regards,
  Enric
END