FROMLIST: arm64: defconfig: Enable options to support panel display for Mediatek Chromebooks
There are some Mediatek based Chromebooks supported in the kernel. Enable the
required config options to have the panel display working on both devices.
This was tested on the ACER Chromebook R13 (MT8173) and the Lenovo
Ideapad Duet (MT8183), but should also enable display support for similar
devices.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Commit-notes:
This is only a resend rebased on top of mainline to fix some trivial
conflicts.
END